feat(SPV-959): Fixed react-refresh plugin issue, moved files
