+++ src/components/ContactAcceptDialog/ContactAcceptDialog.tsx
@@ -5,7 +5,7 @@ import { useState } from 'react';
 
 import { toast } from 'sonner';
 
-import { Button } from '@/components/ui/button.tsx';
+import { Button } from '@/components/ui';
 import {
   Dialog,
   DialogContent,
