+++ src/components/ContactDeleteDialog/ContactDeleteDialog.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 
 import { toast } from 'sonner';
 
-import { Button } from '@/components/ui/button.tsx';
+import { Button } from '@/components/ui';
 import {
   Dialog,
   DialogContent,
