+++ src/components/DataTable/DataTablePagination.tsx
@@ -6,7 +6,7 @@ import {
 } from '@heroicons/react/24/outline';
 import { Table } from '@tanstack/react-table';
 
-import { Button } from '@/components/ui/button.tsx';
+import { Button } from '@/components/ui';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select.tsx';
 
 interface DataTablePaginationProps<TData> {
