+++ src/components/RecordTxDialogAdmin/RecordTxDialogAdmin.tsx
@@ -5,7 +5,7 @@ import React, { useState } from 'react';
 
 import { toast } from 'sonner';
 
-import { Button } from '@/components/ui/button.tsx';
+import { Button } from '@/components/ui';
 import {
   Dialog,
   DialogContent,
