+++ src/components/TransactionsColumns/columns.tsx
@@ -5,7 +5,7 @@ import { ColumnDef } from '@tanstack/react-table';
 import { ArrowUpDown } from 'lucide-react';
 
 import { Badge } from '@/components';
-import { Button } from '@/components/ui/button.tsx';
+import { Button } from '@/components/ui';
 import { getSortDirection } from '@/utils';
 
 export const columns: ColumnDef<Tx>[] = [
