+++ src/components/XPubColumns/XPubsColumns.tsx
@@ -4,9 +4,7 @@ import { ColumnDef } from '@tanstack/react-table';
 
 import { ArrowUpDown } from 'lucide-react';
 
-import { Badge } from '@/components/ui/badge.tsx';
-
-import { Button } from '@/components/ui/button.tsx';
+import { Badge, Button } from '@/components/ui';
 import { getSortDirection } from '@/utils';
 
 export interface XpubsColumns extends XPub {
