+++ src/components/ui/badge/badge.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+import { type VariantProps } from 'class-variance-authority';
+
+import { cn } from '@/lib/utils.ts';
+import { badgeVariants } from '@/components/ui/badge/badgeVariants.tsx';
+
+export interface BadgeProps extends React.HTMLAttributes<HTMLDivElement>, VariantProps<typeof badgeVariants> {}
+
+function Badge({ className, variant, ...props }: BadgeProps) {
+  return <div className={cn(badgeVariants({ variant }), className)} {...props} />;
+}
+
+export { Badge };
