+++ src/components/ui/badge/index.ts
@@ -1,0 +1,2 @@
+export * from './badge.tsx';
+export * from './badgeVariants.tsx';
