+++ src/components/ui/index.ts
@@ -1,7 +1,7 @@
 export * from './accordion.tsx';
 export * from './avatar.tsx';
-export * from './badge.tsx';
-export * from './button.tsx';
+export * from './badge';
+export * from './button';
 export * from './calendar.tsx';
 export * from './card.tsx';
 export * from './dialog.tsx';
