+++ src/contexts/AuthContext/AuthContext.tsx
@@ -1,15 +1,6 @@
-import React, { createContext, useContext, useState } from 'react';
+import React, { createContext, useState } from 'react';
 
-import { useSpvWalletClient } from '@/contexts';
-
-// TODO [react-refresh]: only 1 export is allowed
-// eslint-disable-next-line  react-refresh/only-export-components
-export const enum Role {
-  Admin = 'admin',
-  User = 'user',
-}
-
-export type TRole = Role | null | undefined;
+import { Role, useSpvWalletClient } from '@/contexts';
 
 export interface AuthContext {
   isAuthenticated: boolean;
@@ -19,7 +10,7 @@ export interface AuthContext {
   setLoginKey: React.Dispatch<React.SetStateAction<string>>;
 }
 
-const AuthContext = createContext<AuthContext | null>(null);
+export const AuthContext = createContext<AuthContext | null>(null);
 
 export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
   const { spvWalletClient } = useSpvWalletClient();
@@ -35,14 +26,4 @@ export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
       {children}
     </AuthContext.Provider>
   );
-};
-
-// TODO [react-refresh]: only 1 export is allowed
-// eslint-disable-next-line  react-refresh/only-export-components
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-  return context;
 };
