+++ src/contexts/AuthContext/types.ts
@@ -1,0 +1,6 @@
+export const enum Role {
+  Admin = 'admin',
+  User = 'user',
+}
+
+export type TRole = Role | null | undefined;
