+++ src/contexts/SpvWalletContext/index.ts
@@ -1,0 +1,2 @@
+export * from './SpvWalletContext.tsx';
+export * from './useSpvWalletClient.tsx';
