+++ src/contexts/ThemeProviderContext/index.ts
@@ -1,0 +1,2 @@
+export * from './ThemeProviderContext.tsx';
+export * from './useTheme.tsx';
