+++ src/contexts/index.ts
@@ -1,3 +1,3 @@
-export * from './SpvWalletContext.tsx';
-export * from './ThemeProviderContext.tsx';
-export * from './AuthContext.tsx';
+export * from './SpvWalletContext';
+export * from './ThemeProviderContext';
+export * from './AuthContext';
