+++ src/routes/admin/_admin.contacts.tsx
@@ -25,21 +25,17 @@ import {
 import { useSpvWalletClient } from '@/contexts';
 import { contactsQueryOptions, getContactId, getContactPaymail } from '@/utils';
 
-// TODO [react-refresh]: only 1 export is allowed
-// eslint-disable-next-line  react-refresh/only-export-components
-export const contactsSearchSchema = z.object({
-  createdRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
-  order_by_field: z.string().optional().catch('id'),
-  sort_direction: z.string().optional().catch('desc'),
-  updatedRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
-  id: z.string().optional(),
-  paymail: z.string().optional(),
-  pubKey: z.string().optional(),
-});
-
 export const Route = createFileRoute('/admin/_admin/contacts')({
   component: Contacts,
-  validateSearch: contactsSearchSchema,
+  validateSearch: z.object({
+    createdRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
+    order_by_field: z.string().optional().catch('id'),
+    sort_direction: z.string().optional().catch('desc'),
+    updatedRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
+    id: z.string().optional(),
+    paymail: z.string().optional(),
+    pubKey: z.string().optional(),
+  }),
   loaderDeps: ({ search: { order_by_field, sort_direction, createdRange, updatedRange, id, paymail, pubKey } }) => ({
     order_by_field,
     sort_direction,
