+++ src/routes/admin/_admin.destinations.tsx
@@ -2,22 +2,10 @@ import { createFileRoute, useLoaderData, useNavigate, useSearch } from '@tanstac
 import { useEffect, useState } from 'react';
 import { useDebounce } from 'use-debounce';
 
-import { z } from 'zod';
-
 import { DateRangeFilter, Searchbar, Tabs, TabsContent, TabsList, TabsTrigger, Toaster } from '@/components';
 import { DestinationsTabContent } from '@/components/DestinationsTabContent';
 import { addStatusField, getAddress, getDeletedElements, getLockingScript } from '@/utils';
-
-// TODO [react-refresh]: only 1 export is allowed
-// eslint-disable-next-line  react-refresh/only-export-components
-export const destinationSearchSchema = z.object({
-  lockingScript: z.string().optional().catch(''),
-  address: z.string().optional().catch(''),
-  order_by_field: z.string().optional().catch('id'),
-  sort_direction: z.string().optional().catch('desc'),
-  createdRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
-  updatedRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
-});
+import { destinationSearchSchema } from '@/searchSchemas/destinationSearchSchema.tsx';
 
 export const Route = createFileRoute('/admin/_admin/destinations')({
   component: Destinations,
