+++ src/routes/admin/_admin.paymails.tsx
@@ -21,19 +21,15 @@ import {
 import { useSpvWalletClient } from '@/contexts';
 import { addStatusField, getDeletedElements, paymailsQueryOptions } from '@/utils';
 
-// TODO [react-refresh]: only 1 export is allowed
-// eslint-disable-next-line  react-refresh/only-export-components
-export const paymailsSearchSchema = z.object({
-  order_by_field: z.string().optional().catch('id'),
-  sort_direction: z.string().optional().catch('desc'),
-  xpubId: z.string().optional().catch(''),
-  createdRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
-  updatedRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
-});
-
 export const Route = createFileRoute('/admin/_admin/paymails')({
   component: Paymails,
-  validateSearch: paymailsSearchSchema,
+  validateSearch: z.object({
+    order_by_field: z.string().optional().catch('id'),
+    sort_direction: z.string().optional().catch('desc'),
+    xpubId: z.string().optional().catch(''),
+    createdRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
+    updatedRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
+  }),
   loaderDeps: ({ search: { order_by_field, sort_direction, xpubId, createdRange, updatedRange } }) => ({
     order_by_field,
     sort_direction,
