+++ src/routes/admin/_admin.transactions.tsx
@@ -4,8 +4,6 @@ import { createFileRoute, useSearch } from '@tanstack/react-router';
 import { useState } from 'react';
 import { useDebounce } from 'use-debounce';
 
-import { z } from 'zod';
-
 import {
   Searchbar,
   Tabs,
@@ -18,16 +16,7 @@ import {
 } from '@/components';
 import { useSpvWalletClient } from '@/contexts';
 import { transactionsQueryOptions } from '@/utils';
-
-// TODO [react-refresh]: only 1 export is allowed
-// eslint-disable-next-line  react-refresh/only-export-components
-export const transactionSearchSchema = z.object({
-  order_by_field: z.string().optional().catch('id'),
-  sort_direction: z.string().optional().catch('desc'),
-  blockHeight: z.number().optional().catch(undefined),
-  createdRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
-  updatedRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
-});
+import { transactionSearchSchema } from '@/searchSchemas';
 
 export const Route = createFileRoute('/admin/_admin/transactions')({
   component: Transactions,
