+++ src/searchSchemas/index.ts
@@ -1,0 +1,2 @@
+export * from './destinationSearchSchema.tsx';
+export * from './transactionSearchSchema.tsx';
