+++ src/searchSchemas/transactionSearchSchema.tsx
@@ -1,0 +1,9 @@
+import { z } from 'zod';
+
+export const transactionSearchSchema = z.object({
+  order_by_field: z.string().optional().catch('id'),
+  sort_direction: z.string().optional().catch('desc'),
+  blockHeight: z.number().optional().catch(undefined),
+  createdRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
+  updatedRange: z.object({ from: z.string(), to: z.string() }).optional().catch(undefined),
+});
